--- routes/test.routes.js
@@ -1,7 +1,0 @@
-const express = require('express');
-const testControllers = require('../controllers/test.controllers');
-const router = express.Router();
-
-router.get('/', testControllers.test);
-
-module.exports = router
